messeging deletion 100% now works and fixed some problems with loadin in profile page

+++ Backend/Controllers/MessagingController.cs
@@ -35,7 +35,7 @@ namespace MyApiProject.Controllers
 
         [HttpDelete("DeleteMessage")]
         [Authorize]
-        public async Task<IActionResult> DeleteMessage([FromQuery] int messageId)
+        public async Task<IActionResult> DeleteMessage([FromQuery] long messageId)
         {
             try
             {
