+++ Backend/Interfaces/IMessagingService.cs
@@ -5,6 +5,6 @@ public interface IMessagingService
     Task<int> SendMessageAsync(int senderId, int receiverId, string content);
     Task<List<MessageResponseDto>> GetMessagesAsync(int userId, int contactId, int page, int pageSize);
 
-    Task<bool> DeleteMessageAsync(int messageId, int userId);
+    Task<bool> DeleteMessageAsync(long messageId, int userId);
     Task<int> SendMessageWithValidationAsync(int senderId, MessageDto request);
 }
